Fix interactive blocks and more refactoring
